scripts/requirements.txt
paddleocr>=3.4.0
paddlepaddle>=3.0.0
pdf2image>=1.16.0
Pillow>=9.0.0
